Cleaning model elements
